move set_logger_verbosity and setup_outputdir to common
