+++ text/src/autogluon/text/text_prediction/mx/models.py
@@ -18,7 +18,6 @@ from typing import Tuple
 from packaging import version as py_version
 
 from autogluon.core.scheduler.scheduler_factory import scheduler_factory
-from autogluon.core.utils import set_logger_verbosity
 from sklearn.preprocessing import LabelEncoder
 import mxnet as mx
 from mxnet.util import use_np
@@ -31,6 +30,7 @@ from autogluon_contrib_nlp.utils.misc import grouper, \
 from autogluon_contrib_nlp.utils.parameter import move_to_ctx, clip_grad_global_norm
 
 from autogluon.common.utils.multiprocessing_utils import force_forkserver
+from autogluon.common.utils.miscs import set_logger_verbosity
 from autogluon.core import args, space
 from autogluon.core.utils import in_ipynb, verbosity2loglevel
 from autogluon.core.utils.utils import get_cpu_count, get_gpu_count_mxnet
