+++ vision/src/autogluon/vision/predictor/predictor.py
@@ -15,9 +15,9 @@ except ImportError:
     timm = None
 
 from autogluon.common.utils.multiprocessing_utils import is_fork_enabled
+from autogluon.common.utils.miscs import set_logger_verbosity
 from autogluon.core.constants import MULTICLASS, BINARY, REGRESSION
 from autogluon.core.data.label_cleaner import LabelCleaner
-from autogluon.core.utils import set_logger_verbosity
 from autogluon.core.utils import verbosity2loglevel, get_gpu_count_all
 from autogluon.core.utils.utils import generate_train_test_split
